+++ src/page-objects/administration/CategoryDetail.ts
@@ -1,8 +1,8 @@
 import type { Page, Locator } from '@playwright/test';
 import type { PageObject } from '../../types/PageObject';
-import {CustomFieldLocators} from './modules/CustomFields';
 import { satisfies } from 'compare-versions';
 import { HelperFixtureTypes } from '../../fixtures/HelperFixtures';
+import { getCustomFieldCardLocators } from './modules/CustomFieldCard';
 
 export class CategoryDetail implements PageObject {
     public readonly saveButton: Locator;
@@ -10,7 +10,6 @@ export class CategoryDetail implements PageObject {
     public readonly customFieldCard: Locator;
     public readonly customFieldSetTabs: Locator;
     public readonly customFieldSetTabCustomContent: Locator;
-    public readonly customFieldLocators: CustomFieldLocators;
 
     constructor(public readonly page: Page, public readonly instanceMeta: HelperFixtureTypes['InstanceMeta']) {
         this.saveButton = page.getByRole('button', { name: 'Save' });
@@ -24,7 +23,6 @@ export class CategoryDetail implements PageObject {
 
         this.customFieldSetTabs = this.customFieldCard.locator('.sw-tabs-item');
         this.customFieldSetTabCustomContent = this.customFieldCard.locator('.sw-tabs__custom-content');
-        this.customFieldLocators = new CustomFieldLocators(page, instanceMeta);
     }
 
     async getCustomFieldSetCardContentByName(customFieldSetName: string): Promise<Record<string, Locator>> {
@@ -48,7 +46,7 @@ export class CategoryDetail implements PageObject {
         return `#/sw/category/index/${categoryUuid}/base`
     }
 
-    async customFields(customFieldSetName: string, customFieldTextName: string) {
-        return await this.customFieldLocators.getLocators(customFieldSetName, customFieldTextName);
+    async getCustomFieldCardLocators(customFieldSetName: string, customFieldTextName: string) {
+        return getCustomFieldCardLocators(this.page, customFieldSetName, customFieldTextName, this.instanceMeta);
     }
 }
